requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-cors-headers==3.4.0
django-environ==0.4.5
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
flake8==3.8.3
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.4
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
qrcode==6.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
zipp==3.1.0
